renamed effect mechanics to respect nomenclature
