--- GW2EIEvtcParser/EIData/Mechanics/EffectMechanics/EffectSrcMechanic.cs
@@ -1,27 +1,0 @@
-﻿using System.Linq;
-using GW2EIEvtcParser.ParsedData;
-
-namespace GW2EIEvtcParser.EIData
-{
-
-    internal abstract class EffectSrcMechanic : EffectMechanic
-    {
-
-        protected override AgentItem GetAgentItem(EffectEvent effectEvt, AgentData agentData)
-        {
-            if (effectEvt.Src == ParserHelper._unknownAgent || effectEvt.Src.IsSpecy(ArcDPSEnums.NonIdentifiedAgent))
-            {
-                return agentData.GetNPCsByID(ArcDPSEnums.TrashID.Environment).FirstOrDefault();
-            }
-            return effectEvt.Src;
-        }
-
-        public EffectSrcMechanic(string effectGUID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : this(new string[] { effectGUID }, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
-        {
-        }
-
-        public EffectSrcMechanic(string[] effectGUIDs, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : base(effectGUIDs, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
-        {
-        }
-    }
-}
