--- GW2EIEvtcParser/EIData/Mechanics/EffectMechanics/Enemy/EffectSrcEnemyMechanic.cs
@@ -1,25 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using GW2EIEvtcParser.ParsedData;
-
-namespace GW2EIEvtcParser.EIData
-{
-
-    internal class EffectSrcEnemyMechanic : EffectSrcMechanic
-    {
-
-        public EffectSrcEnemyMechanic(string effectGUID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : this(new string[] { effectGUID }, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
-        {
-        }
-
-        public EffectSrcEnemyMechanic(string[] effectGUIDs, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : base(effectGUIDs, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
-        {
-            IsEnemyMechanic = true;
-        }
-
-        internal override void CheckMechanic(ParsedEvtcLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<int, AbstractSingleActor> regroupedMobs)
-        {
-            EnemyChecker(log, mechanicLogs, regroupedMobs);
-        }
-    }
-}
