+++ GW2EIEvtcParser/EIData/Mechanics/EffectMechanics/Friendly/PlayerDstEffectMechanic.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using GW2EIEvtcParser.ParsedData;
+
+namespace GW2EIEvtcParser.EIData
+{
+
+    internal class PlayerDstEffectMechanic : DstEffectMechanic
+    {
+
+        public PlayerDstEffectMechanic(string effectGUID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : this(new string[] { effectGUID }, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+
+        public PlayerDstEffectMechanic(string[] effectGUIDs, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, EffectChecker condition = null) : base(effectGUIDs, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+
+        internal override void CheckMechanic(ParsedEvtcLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<int, AbstractSingleActor> regroupedMobs)
+        {
+            PlayerChecker(log, mechanicLogs);
+        }
+    }
+}
